refactor: move current_active_wips and all_wips_active to mainnet_validations
