+++ rad/src/lib.rs
@@ -16,7 +16,7 @@ use crate::{
 };
 use witnet_data_structures::{
     chain::{RADAggregate, RADRequest, RADRetrieve, RADTally, RADType},
-    mainnet_validations::{wip_info, ActiveWips},
+    mainnet_validations::{current_active_wips, ActiveWips},
     radon_report::{RadonReport, ReportContext, RetrievalMetadata, Stage, TallyMetaData},
 };
 
@@ -40,24 +40,6 @@ pub struct RADRequestExecutionReport {
     pub retrieve: Vec<RadonReport<RadonTypes>>,
     /// Report about aggregation of reports (reveals, actually).
     pub tally: RadonReport<RadonTypes>,
-}
-
-/// Auxiliary function that returns the current active wips for using in RADON
-/// It is only used for testing or for external libraries, so we set epoch to MAX
-pub fn current_active_wips() -> ActiveWips {
-    ActiveWips {
-        active_wips: wip_info(),
-        block_epoch: u32::MAX,
-    }
-}
-
-/// Auxiliary function that returns the current active wips and the WIPs in voting process as actived
-/// It is only used for testing
-pub fn all_wips_active() -> ActiveWips {
-    let mut h = current_active_wips();
-    h.active_wips.insert("WIP0017-0018-0019".to_string(), 0);
-
-    h
 }
 
 /// Executes a data request locally.
